+++ frontend/src/App.jsx
@@ -1,10 +1,11 @@
 import content from "./data/content.json";
 import { ModernHeader } from "./components/ModernHeader";
 import { ModernWhatIDo } from "./components/ModernWhatIDo";
-import { Skills } from "./components/Skills";
 import { ModernProjects } from "./components/ModernProjects";
 import { AboutMe } from "./components/AboutMe";
+import { Contact } from "./components/Contact";
 import { Footer } from "./components/Footer";
+import { ScrollToTop } from "./components/ScrollToTop";
 
 export const App = () => {
   return (
@@ -14,15 +15,12 @@ export const App = () => {
         email={content.headerSection.email}
         mailtoText={content.headerSection.mailtoText}
       />
-
       <ModernWhatIDo />
-
-
       <ModernProjects />
-
       <AboutMe />
-
+      <Contact />
       <Footer />
+      <ScrollToTop />
     </>
   );
 };
